--- src/MOI/solve.jl
@@ -1,5 +1,0 @@
-function MOI.optimize!(m::CplexSolverInstance)
-    # start = time()
-    optimize!(m.inner)
-    # m.solvetime = time() - start
-end
