+++ src/cpx_variables.jl
