+++ src/moi_constraints.jl
@@ -1,4 +1,3 @@
-include("constraints_cpx.jl")
 
 constraint_storage(m::CplexSolverInstance, func::MOI.AbstractScalarFunction, set::MOI.AbstractSet) = constraint_storage(m, typeof(func), typeof(set))
 constraint_storage(m::CplexSolverInstance, ::Type{Linear}, ::Type{LE}) = m.constraint_mapping.less_than
@@ -10,6 +9,10 @@ constraint_storage(m::CplexSolverInstance, ::Type{MOI.SingleVariable}, ::Type{GE
 constraint_storage(m::CplexSolverInstance, ::Type{MOI.SingleVariable}, ::Type{EQ}) = m.constraint_mapping.fixed_variables
 constraint_storage(m::CplexSolverInstance, ::Type{MOI.SingleVariable}, ::Type{MOI.Interval{Float64}}) = m.constraint_mapping.interval_variables
 
+function constraint_storage_value(m::CplexSolverInstance, c::MOI.ConstraintReference{F, S}) where F where S
+    dict = constraint_storage(m, F, S)
+    return dict[c]
+end
 """
     Set bounds on variable
 """
@@ -39,49 +42,56 @@ function setvariablebound!(m::CplexSolverInstance, v::MOI.SingleVariable, set::M
     setvariablebound!(m, getcols(m, v), set.lower, Cchar('L'))
 end
 
-function MOI.getattribute(m::CplexSolverInstance, ::MOI.ConstraintSet, c::MOI.ConstraintReference{MOI.SingleVariable,MOI.LessThan{Float64}})
-    vref = m.constraint_mapping[c]::MOI.VariableReference
-    ub = cpx_get_variable_upperbound(m.inner, getcol(m, vref))
-    return MOI.LessThan{Float64}(ub)
+function MOI.getattribute(m::CplexSolverInstance, ::MOI.ConstraintSet, c::MOI.ConstraintReference{MOI.SingleVariable,T}) where T <: Union{LE, GE, EQ}
+    varref = constraint_storage_value(m, c)
+    val = get_bound_value(m, varref, T)
+    return T(val)
 end
-function MOI.getattribute(m::CplexSolverInstance, ::MOI.ConstraintSet, c::MOI.ConstraintReference{MOI.SingleVariable,MOI.GreaterThan{Float64}})
-    vref = m.constraint_mapping[c]::MOI.VariableReference
-    lb = cpx_get_variable_lowerbound(m.inner, getcol(m, vref))
-    return MOI.GreaterThan{Float64}(lb)
+
+function get_bound_value(m::CplexSolverInstance, v::MOI.VariableReference, ::Type{LE})
+    cpx_get_variable_upperbound(m.inner, getcol(m, v))
 end
-function MOI.getattribute(m::CplexSolverInstance, ::MOI.ConstraintSet, c::MOI.ConstraintReference{MOI.SingleVariable,MOI.EqualTo{Float64}})
-    vref = m.constraint_mapping[c]::MOI.VariableReference
-    lb = cpx_get_variable_lowerbound(m.inner, getcol(m, vref))
-    return MOI.EqualTo{Float64}(lb)
+function get_bound_value(m::CplexSolverInstance, v::MOI.VariableReference, ::Type{GE})
+    cpx_get_variable_lowerbound(m.inner, getcol(m, v))
+end
+function get_bound_value(m::CplexSolverInstance, v::MOI.VariableReference, ::Type{EQ})
+    cpx_get_variable_lowerbound(m.inner, getcol(m, v))
 end
 function MOI.getattribute(m::CplexSolverInstance, ::MOI.ConstraintSet, c::MOI.ConstraintReference{MOI.SingleVariable,MOI.Interval{Float64}})
-    vref = m.constraint_mapping[c]::MOI.VariableReference
-    lb = cpx_get_variable_lowerbound(m.inner, getcol(m, vref))
-    ub = cpx_get_variable_upperbound(m.inner, getcol(m, vref))
-    return MOI.Interval{Float64}(lb, ub)
+    varref = constraint_storage_value(m, c)
+    lb = cpx_get_variable_lowerbound(m.inner, getcol(m, varref))
+    ub = cpx_get_variable_upperbound(m.inner, getcol(m, varref))
+    return Interval{Float64}(lb, ub)
 end
+
+MOI.cangetattribute(m::CplexSolverInstance, ::MOI.ConstraintSet, c::MOI.ConstraintReference{MOI.SingleVariable,T}) where T <: Union{LE, GE, EQ, MOI.Interval{Float64}} = true
 
 function MOI.getattribute(m::CplexSolverInstance, ::MOI.ConstraintFunction, c::MOI.ConstraintReference{MOI.SingleVariable,T}) where T
-    vref = m.constraint_mapping[c]::MOI.VariableReference
+    vref = constraint_storage_value(m, c)
     return MOI.SingleVariable(vref)
 end
+MOI.cangetattribute(m::CplexSolverInstance, ::MOI.ConstraintFunction, c::MOI.ConstraintReference{MOI.SingleVariable,T}) where T <: Union{LE, GE, EQ, MOI.Interval{Float64}} = true
 
+
+"""
+    Add linear constraints
+"""
 function MOI.addconstraint!(m::CplexSolverInstance, func::Linear, set::T) where T
     addlinearconstraint!(m, func, set)
     m.last_constraint_reference += 1
     ref = MOI.ConstraintReference{Linear, T}(m.last_constraint_reference)
     dict = constraint_storage(m, func, set)
-    dict[ref] = num_constr(m.inner)
+    dict[ref] = cpx_number_constraints(m.inner)
     return ref
 end
 
-function addlinearconstraint!(m::CplexSolverInstance, func::Linear, set::MOI.LessThan{Float64})
+function addlinearconstraint!(m::CplexSolverInstance, func::Linear, set::LE)
     addlinearconstraint!(m, func, Cchar('L'), set.upper)
 end
-function addlinearconstraint!(m::CplexSolverInstance, func::Linear, set::MOI.GreaterThan{Float64})
+function addlinearconstraint!(m::CplexSolverInstance, func::Linear, set::GE)
     addlinearconstraint!(m, func, Cchar('G'), set.lower)
 end
-function addlinearconstraint!(m::CplexSolverInstance, func::Linear, set::MOI.EqualTo{Float64})
+function addlinearconstraint!(m::CplexSolverInstance, func::Linear, set::EQ)
     addlinearconstraint!(m, func, Cchar('E'), set.value)
 end
 
@@ -95,28 +105,14 @@ end
 function MOI.getattribute(m::CplexSolverInstance, ::MOI.NumberOfConstraints{F, S}) where F where S
     length(constraint_storage(m, F, S))
 end
+MOI.cangetattribute(m::CplexSolverInstance, ::MOI.NumberOfConstraints{F, S}) where F where S = true
 
-# function MOI.getattribute(m::CplexSolverInstance, ::ConstraintSet, c::ConstraintReference{ScalarAffineFunction{Float64},LessThan{Float64}})
-#     row = m.constraint_mapping[c]::Int
-#     ub = cpx_get_variable_upperbound(m.inner, getcol(m, vref))
-#     return LessThan{Float64}(ub)
-# end
-# function MOI.getattribute(m::CplexSolverInstance, ::ConstraintSet, c::ConstraintReference{SingleVariable,GreaterThan{Float64}})
-#     row = m.constraint_mapping[c]::Int
-#     lb = cpx_get_variable_lowerbound(m.inner, getcol(m, vref))
-#     return GreaterThan{Float64}(lb)
-# end
-# function MOI.getattribute(m::CplexSolverInstance, ::ConstraintSet, c::ConstraintReference{SingleVariable,EqualsTo{Float64}})
-#     row = m.constraint_mapping[c]::Int
-#     lb = cpx_get_variable_lowerbound(m.inner, getcol(m, vref))
-#     return EqualsTo{Float64}(lb)
-# end
-# function MOI.getattribute(m::CplexSolverInstance, ::ConstraintSet, c::ConstraintReference{SingleVariable,Interval{Float64}})
-#     row = m.constraint_mapping[c]::Int
-#     lb = cpx_get_variable_lowerbound(m.inner, getcol(m, vref))
-#     ub = cpx_get_variable_upperbound(m.inner, getcol(m, vref))
-#     return Interval{Float64}(lb, ub)
-# end
+function MOI.getattribute(m::CplexSolverInstance, ::MOI.ConstraintSet, c::MOI.ConstraintReference{Linear,T}) where T <: Union{LE, GE, EQ}
+    row = constraint_storage_value(m, c)::Int
+    rhs = cpx_get_rhs(m.inner, row)
+    T(rhs)
+end
+MOI.cangetattribute(m::CplexSolverInstance, ::MOI.ConstraintSet, ::MOI.ConstraintReference{Linear,T}) where T <: Union{LE, GE, EQ} = true
 #
 # function MOI.getattribute(m::CplexSolverInstance, ::ConstraintFunction, c::ConstraintReference{ScalarAffineFunction{Float64},T}) where T <: Union{LessThan{Float64}, GreaterThan{Float64}, EqualsTo{Float64}}
 #     row = m.constraint_mapping[c]::Int
