+++ src/old/cpx_vars.jl
