+++ test/old/lp_02.jl
